Move Hystrix code from Admin to a decdicated file
