+++ source/common/stats/hystrix.h
@@ -2,6 +2,8 @@
 #include <memory>
 #include <vector>
 
+#include "envoy/server/admin.h"
+#include "envoy/server/instance.h"
 #include "envoy/stats/stats.h"
 
 namespace Envoy {
@@ -100,6 +102,61 @@ private:
 };
 
 typedef std::unique_ptr<Hystrix> HystrixPtr;
+
+/**
+ * This class contains data which will be sent from admin filter to a hystrix_event_stream handler
+ * and build a class which contains the relevant data.
+ */
+class HystrixHandlerInfoImpl : public Server::HandlerInfo {
+public:
+  HystrixHandlerInfoImpl(Http::StreamDecoderFilterCallbacks* callbacks)
+      : stats_(new Stats::Hystrix()), data_timer_(nullptr), ping_timer_(nullptr),
+        callbacks_(callbacks) {}
+  virtual ~HystrixHandlerInfoImpl(){};
+  virtual void Destroy();
+
+  /**
+   * HystrixHandlerInfoImpl includes statistics for hystrix API, timers for build (and send) data
+   * and keep alive messages and the handler's callback
+   */
+  Stats::HystrixPtr stats_;
+  Event::TimerPtr data_timer_;
+  Event::TimerPtr ping_timer_;
+  Http::StreamDecoderFilterCallbacks* callbacks_{};
+};
+
+/**
+ * Convert statistics from envoy format to hystrix format and prepare them and writes them to the
+ * appropriate socket
+ */
+class HystrixHandler {
+public:
+  static void HandleEventStream(HystrixHandlerInfoImpl* hystrix_handler_info,
+                                Server::Instance& server);
+  /**
+   * Update counter and set values of upstream_rq statistics
+   * @param hystrix_handler_info is the data which is received in the hystrix handler from the admin
+   * filter (callback, timers, statistics)
+   * @param server contains envoy statistics
+   */
+  static void updateHystrixRollingWindow(HystrixHandlerInfoImpl* hystrix_handler_info,
+                                         Server::Instance* server);
+  /**
+   * Builds a buffer of envoy statistics which will be sent to hystrix dashboard according to
+   * hystrix API
+   * @param hystrix_handler_info is the data which is received in the hystrix handler from the admin
+   * filter (callback, timers, statistics)
+   * @param server contains envoy statistics*
+   */
+  static void prepareAndSendHystrixStream(HystrixHandlerInfoImpl* hystrix_handler_info,
+                                          Server::Instance* server);
+  /**
+   * Sends a keep alive (ping) message to hystrix dashboard
+   * @param hystrix_handler_info is the data which is received in the hystrix handler from the admin
+   * filter (callback, timers, statistics)
+   */
+  static void sendKeepAlivePing(HystrixHandlerInfoImpl* hystrix_handler_info);
+};
 
 } // namespace Stats
 } // namespace Envoy
