+++ source/server/http/admin.cc
@@ -590,24 +590,9 @@ Http::Code AdminImpl::handlerHystrixEventStream(const std::string&,
       Http::Headers::get().AccessControlAllowOriginValue.All);
   response_headers.insertNoChunks().value().setReference("0");
 
-  HystrixHandlerInfoImpl& hystrix_handler_info = dynamic_cast<HystrixHandlerInfoImpl&>(handler_info);
-
-  // start streaming
-  hystrix_handler_info.data_timer_ = hystrix_handler_info.callbacks_->dispatcher().createTimer(
-      [this, &hystrix_handler_info]() -> void {
-        HystrixHandler::prepareAndSendHystrixStream(&hystrix_handler_info, server_);
-      });
-  hystrix_handler_info.data_timer_->enableTimer(
-      std::chrono::milliseconds(Stats::Hystrix::GetRollingWindowIntervalInMs()));
-
-  // start keep alive ping
-  hystrix_handler_info.ping_timer_ =
-      hystrix_handler_info.callbacks_->dispatcher().createTimer([&hystrix_handler_info]() -> void {
-        HystrixHandler::sendKeepAlivePing(&hystrix_handler_info);
-      });
-
-  hystrix_handler_info.ping_timer_->enableTimer(
-      std::chrono::milliseconds(Stats::Hystrix::GetPingIntervalInMs()));
+  Stats::HystrixHandlerInfoImpl& hystrix_handler_info =
+      dynamic_cast<Stats::HystrixHandlerInfoImpl&>(handler_info);
+  Stats::HystrixHandler::HandleEventStream(&hystrix_handler_info, server_);
 
   ENVOY_LOG(debug, "start sending data to hystrix dashboard on port {}",
             hystrix_handler_info.callbacks_->connection()->localAddress()->asString());
@@ -627,7 +612,7 @@ void AdminFilter::onComplete() {
     handler_info_ = std::make_unique<HandlerInfoImpl>();
     code = parent_.runCallback(path, *header_map, response, *handler_info_);
   } else {
-    handler_info_ = std::make_unique<HystrixHandlerInfoImpl>(callbacks_);
+    handler_info_ = std::make_unique<Stats::HystrixHandlerInfoImpl>(callbacks_);
     code = parent_.runCallback(path, *header_map, response, *handler_info_);
     end_stream = false;
   }
@@ -844,61 +829,6 @@ bool AdminImpl::removeHandler(const std::string& prefix) {
     return true;
   }
   return false;
-}
-
-void HystrixHandlerInfoImpl::Destroy() {
-  if (data_timer_) {
-    data_timer_->disableTimer();
-    data_timer_.reset();
-  }
-  if (ping_timer_) {
-    ping_timer_->disableTimer();
-    ping_timer_.reset();
-  }
-}
-
-void HystrixHandler::updateHystrixRollingWindow(HystrixHandlerInfoImpl* hystrix_handler_info,
-                                                Server::Instance& server) {
-  hystrix_handler_info->stats_->incCounter();
-  for (auto& cluster : server.clusterManager().clusters()) {
-    hystrix_handler_info->stats_->updateRollingWindowMap(server.stats(),
-                                                         cluster.second.get().info()->name());
-  }
-}
-
-void HystrixHandler::prepareAndSendHystrixStream(HystrixHandlerInfoImpl* hystrix_handler_info,
-                                                 Server::Instance& server) {
-  updateHystrixRollingWindow(hystrix_handler_info, server);
-  std::stringstream ss;
-  for (auto& cluster : server.clusterManager().clusters()) {
-    hystrix_handler_info->stats_->getClusterStats(
-        ss, cluster.second.get().info()->name(),
-        cluster.second.get()
-            .info()
-            ->resourceManager(Upstream::ResourcePriority::Default)
-            .pendingRequests()
-            .max(),
-        server.stats()
-            .gauge("cluster." + cluster.second.get().info()->name() + ".membership_total")
-            .value());
-  }
-  Buffer::OwnedImpl data;
-  data.add(ss.str());
-  hystrix_handler_info->callbacks_->encodeData(data, false);
-
-  // restart timer
-  hystrix_handler_info->data_timer_->enableTimer(
-      std::chrono::milliseconds(Stats::Hystrix::GetRollingWindowIntervalInMs()));
-}
-
-void HystrixHandler::sendKeepAlivePing(HystrixHandlerInfoImpl* hystrix_handler_info) {
-  Buffer::OwnedImpl data;
-  data.add(":\n\n");
-  hystrix_handler_info->callbacks_->encodeData(data, false);
-
-  // restart timer
-  hystrix_handler_info->ping_timer_->enableTimer(
-      std::chrono::milliseconds(Stats::Hystrix::GetPingIntervalInMs()));
 }
 
 } // namespace Server
